pyproject.toml
[project]
name = "my-quant-v2"
version = "2.0.0"
description = "YAML-driven quantitative trading backtest framework built on Backtrader"
authors = [{name = "Anthony Yuan"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "backtrader~=1.9.78.123",
    "yfinance~=0.2.49",
    "pyyaml>=6.0.2",
    "jinja2~=3.1.4",
    "matplotlib~=3.9.2",
    "numpy==2.1.3",
    "python-dateutil~=2.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
